+++ docs/release-notes.md
@@ -7,6 +7,9 @@
   * Updated for the 'Force Off' gamepad mode added in Stardew Valley 1.4.0.1.
   * Fixed compatibility issue with Arch Linux.
 
+* For modders:
+  * Added asset propagation for grass textures.
+
 * For the web UI:
   * If a JSON validator upload can't be saved to Pastebin (e.g. due to rate limits), it's now uploaded to Amazon S3 instead. Files uploaded to S3 expire after one month.
   * Updated the JSON validator for Content Patcher 1.10.0.